Skillbox/Python_django_diploma_dpo/diplom/app_basket/serializers.py
from .models import Cart
from rest_framework import serializers


class CartSerializer(serializers.ModelSerializer):

    class Meta:
        Depth = 1
        model = Cart
        fields = '__all__'
