Skillbox/Python_django_diploma_dpo/requirements.txt
Django~=4.1.5
pip~=20.0.2
distro~=1.4.0
wheel~=0.34.2
cryptography~=2.8
Pillow~=7.0.0
docutils~=0.19
Jinja2~=3.1.2
pytz~=2022.7.1
PyYAML~=5.3.1
asgiref~=3.6.0
sqlparse~=0.2.4
colorama~=0.4.3
numpy~=1.17.4
requests~=2.22.0
setuptools~=45.2.0
uritemplate~=4.1.1
coreapi~=2.3.3
coreschema~=0.0.4
djangorestframework~=3.14.0
django_debug_toolbar==3.8.1